+++ service/controller/v1/prometheus/scrapeconfig_test_vars.go
@@ -996,6 +996,11 @@ var (
 				TargetLabel: ProviderLabel,
 				Replacement: "aws-test",
 			},
+			{
+				Action:       ActionDrop,
+				SourceLabels: model.LabelNames{MetricNameLabel},
+				Regex:        relabel.MustNewRegexp(`(nginx_ingress_controller_request_duration_seconds_bucket|nginx_ingress_controller_response_size_bucket|nginx_ingress_controller_request_size_bucket|nginx_ingress_controller_response_duration_seconds_bucket|nginx_ingress_controller_bytes_sent_bucket)`),
+			},
 		},
 	}
 	TestConfigOneKubeStateManagedApp = config.ScrapeConfig{